requirements.txt
Flask
pandas
joblib
scikit-learn
gunicorn
google-cloud-logging==2.7.0
#in more detailes
botocore>1.23.41,<1.32.0
cement==2.8.2
colorama>=0.2.5,<0.4.4 # use the same range that 'docker-compose' uses
pathspec==0.10.1
python-dateutil>=2.1,<3.0.0 # use the same range that 'botocore' uses
requests>=2.31
setuptools >= 20.0
semantic_version == 2.8.5
six>=1.11.0,<1.17.0
termcolor == 1.1.0
wcwidth>=0.1.7,<0.2.0
PyYAML>=5.3.1,<6.1 # use the same range that 'aws-cli' uses. This is also compatible with 'docker-compose'
urllib3>=1.26.5,<2 #1.26.5 fix CVE-2021-33503
